Fix: social art
Writes directly to dist directory. I strongly suspect this is a bandaid and it's really not at all clear to me why the files aren't moved from staging to dist with the previous code. Anyway it's all the more complicated to test this because social images require a public fully qualified URL to even validate