Remove `onigurama` dependency
Prep for further Node upgrade(s): this package fails to build in later versions.

From memory: `onigurama` is associated with `shiki`. Trying to discern past-me’s intent in having it explicitly installed… I figured it must be a peer dependency. But I don’t see any warnings upon its removal. It seems more likely I’d installed it in the whack-a-mole game trying to figure out which packages to mark “external” in Snowpack…