Replace `require` calls with `import` for various `mdx`-related dependencies…
It’s utterly unclear (without digging deeper) why some of these needed to be marked “external” and others did not. I’d hunch that it’s a difference between some which are properly ESM vs some which are either CJS-only or have incorrectly packaged ESM. But it’s just a hunch. In any case… this works FOR NOW.

Hoping I can do the same with `fela`-related stuff, but that proved more problematic on first attempt.